+++ test/test.dart
@@ -4,13 +4,11 @@ void main() async {
   var hetu = Hetu();
   await hetu.init();
   await hetu.eval(r'''
-  class Super {
-    var name = 'super'
-  }
 
   fun main {
-    var a = Super()
-    print(a.name)
+    var a: fun(any) -> any = fun(n: num) -> num { return n + 1 }
+
+    print(a.rtType)
   }
 
       ''', invokeFunc: 'main');